chore: Update Docker Compose network and volume names

+++ backend/src/users/users.service.ts
@@ -14,6 +14,10 @@ import getMessages from '../lang/getMessages';
 import { Request } from 'express';
 import { REQUEST } from '@nestjs/core';
 import Lang from '../lang/lang.type';
+import { DEFAULT_LANG } from '../lang';
+import { AiApiService } from '../ai-api/ai-api.service';
+import { FileStorageService } from '../file-storage/file-storage.service';
+import { USER } from '../common/constants';
 
 @Injectable({
   scope: Scope.REQUEST,
@@ -25,10 +29,10 @@ export class UsersService {
     private readonly userRepository: Repository<UserEntity>,
     @Inject(REQUEST) private readonly request: Request,
     private dataSource: DataSource,
+    private IAApiService: AiApiService,
+    private fileStorageService: FileStorageService,
   ) {
-    const lang =
-      this.request.headers['accept-language'] ||
-      this.request['preferredLanguage'];
+    const lang = this.request?.['preferredLanguage'] || DEFAULT_LANG;
     this.messages = getMessages(lang);
   }
 
@@ -45,9 +49,23 @@ export class UsersService {
       if (await this.userAlreadyExists(user)) {
         throw new ConflictException(this.messages.USER.ALREADY_EXISTS);
       }
-      const newUser = this.userRepository.create({
+      let profileImageKey: string = null;
+      if (user.profileImage) {
+        const { storageKey } = await this.fileStorageService.uploadFile(
+          user.profileImage,
+          {
+            key: USER.STORAGE_KEY_PATH.PROFILE_IMAGES(user.username),
+          },
+        );
+        profileImageKey = storageKey;
+      }
+
+      const createUserData = {
         ...user,
-      });
+      };
+      delete createUserData.profileImage;
+      createUserData['profileImageKey'] = profileImageKey;
+      const newUser = this.userRepository.create(createUserData);
       // return await this.userRepository.save(newUser);
       const data = await queryRunner.manager.save(newUser);
       await queryRunner.commitTransaction();
@@ -69,30 +87,74 @@ export class UsersService {
     return await this.userRepository.find();
   }
 
-  async findOne(id: string): Promise<UserEntity> {
+  async findOne(id: number) {
+    // let prompt = '';
+    // const instruction = `Instruction: always respond with a joke at the end of the conversation
+
+    //   Message: hi there, how are you?`;
+    // return await this.IAApiService.chatCompletion({
+    //   prompt: instruction,
+    // });
     const user = await this.userRepository.findOne({
       where: { id },
+      select: [
+        'id',
+        'username',
+        'email',
+        'firstName',
+        'lastName',
+        'profileImageKey',
+      ],
     });
+
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return user;
+    return {
+      ...user,
+    };
   }
 
-  async update(id: string, userDto: UserDto): Promise<UserEntity> {
+  async update(id: number, userDto: UserDto): Promise<UserEntity> {
     delete userDto.password; // Don't update the password here
     const user = await this.findOne(id);
+
+    if (userDto.username === user.username) {
+      delete userDto.username;
+      delete userDto.username;
+    }
     const updatedUserData = {
       ...user,
       ...userDto,
+      id: +id,
     };
+    for (const key in updatedUserData) {
+      if (updatedUserData[key] === 'null') {
+        updatedUserData[key] = null;
+      }
+    }
+    let profileImageUrl: string = null;
+    if (userDto.profileImage) {
+      const { storageKey } = await this.fileStorageService.uploadFile(
+        userDto.profileImage,
+        {
+          key: USER.STORAGE_KEY_PATH.PROFILE_IMAGES(user.username),
+        },
+      );
+
+      updatedUserData.profileImageKey = storageKey;
+      profileImageUrl = await this.fileStorageService.getSignedUrl(storageKey);
+    }
+    delete updatedUserData.profileImage;
+
     const updatedUser = await this.userRepository.save(updatedUserData);
+    updatedUser.profileImageUrl = profileImageUrl;
     delete updatedUser.password;
 
     return updatedUser;
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: number): Promise<void> {
     await this.findOne(id);
     await this.userRepository.delete(id);
   }
@@ -104,15 +166,32 @@ export class UsersService {
   }
 
   async findByEmail(email: string): Promise<UserEntity> {
-    return await this.userRepository.findOneOrFail({
+    return await this.userRepository.findOne({
       where: { email },
     });
   }
 
   async deleteByEmail(email: string): Promise<void> {
-    const user = await this.userRepository.findOneOrFail({
+    const user = await this.userRepository.findOne({
       where: { email },
     });
-    await this.userRepository.delete(user.id);
+
+    if (user) {
+      await this.userRepository.delete(user.id);
+    }
+  }
+
+  async createUserIfNotExists(user: UserDto): Promise<UserEntity> {
+    // if (await this.userAlreadyExists(user)) {
+    //   return null;
+    // }
+    let userEntity = await this.userRepository.findOne({
+      where: [{ email: user.email }, { username: user.username }],
+    });
+    if (!userEntity) {
+      userEntity = await this.create(user);
+    }
+
+    return userEntity;
   }
 }
